add es01 logs
this commit adds functionality to screen es01 logs in kibana.

to use, start the project with 'make elk-up'. this makes sure that es01 logs are redirected into the correct folder.

the folder is mounted in docker-compose.yml and filebeat and logstash are configured to take both es01 and app logs